requirements.txt
astroid==3.2.4
black==24.8.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
distlib==0.3.8
filelock==3.15.4
flake8==7.1.1
identify==2.6.0
idna==3.8
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pre-commit==3.8.0
pycodestyle==2.12.1
pyflakes==3.2.0
pylint==3.2.7
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
tomlkit==0.13.2
urllib3==2.2.2
virtualenv==20.26.3
